FINALLY!
Praise God! After weeks the collision detection is finally done!

This time I found not logical else if's statements in the SAT function. It was two else if's that had to be changed to ifs. The if statement was also missing a condition for null.

Moved the functions to another file to make the Game file more cleaner.

And Again Thanks to Qixotl LFC from Youtube. He was a heavy influence to my code when it came to figuring out the SAT algorithm/ collision detection.

His video: https://www.youtube.com/watch?v=MvlhMEE9zuc